+++ tests/test_evaluation_metrics.py
@@ -1,25 +1,27 @@
-from wsknn.evaluate.metrics import mrr_func, recall_func, precision_func
+from wsknn.evaluate.scores.scores import mrr_func, recall_func, precision_func
 
 
 RELEVANT_ITEMS = list('abcdefgh')
 RELEVANT_ITEMS_RECALL = list('abkl')
-PREDICTIONS = list('adk')
+RECOMMENDATIONS = list('adk')
 
+
+# TEST SCORES
 
 def test_mrr():
-    score = mrr_func(preds=PREDICTIONS, rel_items=RELEVANT_ITEMS)
+    score = mrr_func(recommendations=RECOMMENDATIONS, relevant_items=RELEVANT_ITEMS)
     # 1.0
     assert score == 1
 
 
 def test_precision():
-    score = precision_func(preds=PREDICTIONS, rel_items=RELEVANT_ITEMS, k=len(PREDICTIONS))
+    score = precision_func(recommendations=RECOMMENDATIONS, relevant_items=RELEVANT_ITEMS)
     # 0.(6)
     assert score > 0.65
     assert score < 0.67
 
 
 def test_recall():
-    score = recall_func(preds=PREDICTIONS, rel_items=RELEVANT_ITEMS_RECALL)
+    score = recall_func(recommendations=RECOMMENDATIONS, relevant_items=RELEVANT_ITEMS_RECALL)
     # 0.5
     assert score == 0.5